Fix build warning for lwip_isdigit/isxdigit/islower/isspace
lwip_isdigit/isxdigit/islower/isspace takes int as argument, so
explicitly cast to int to silence below build warnings.

src/core/ipv4/ip4_addr.c: In function 'ip4addr_aton':
src/core/ipv4/ip4_addr.c:160:5: warning: array subscript has type 'char' [-Wchar-subscripts]
     if (!lwip_isdigit(c)) {
     ^
src/core/ipv4/ip4_addr.c:175:7: warning: array subscript has type 'char' [-Wchar-subscripts]
       if (lwip_isdigit(c)) {
       ^
src/core/ipv4/ip4_addr.c:178:7: warning: array subscript has type 'char' [-Wchar-subscripts]
       } else if (base == 16 && lwip_isxdigit(c)) {
       ^
src/core/ipv4/ip4_addr.c:179:9: warning: array subscript has type 'char' [-Wchar-subscripts]
         val = (val << 4) | (u32_t)(c + 10 - (lwip_islower(c) ? 'a' : 'A'));
         ^
src/core/ipv4/ip4_addr.c:204:3: warning: array subscript has type 'char' [-Wchar-subscripts]
   if (c != '\0' && !lwip_isspace(c)) {
   ^


# Conflicts:
#	src/include/lwip/arch.h

+++ src/include/lwip/arch.h
@@ -226,10 +226,10 @@ typedef int ssize_t;
 #define lwip_toupper(c)           (lwip_islower(c) ? (c) - 'a' + 'A' : c)
 #else
 #include <ctype.h>
-#define lwip_isdigit(c)           isdigit(c)
-#define lwip_isxdigit(c)          isxdigit(c)
-#define lwip_islower(c)           islower(c)
-#define lwip_isspace(c)           isspace(c)
+#define lwip_isdigit(c)           isdigit((int)c)
+#define lwip_isxdigit(c)          isxdigit((int)c)
+#define lwip_islower(c)           islower((int)c)
+#define lwip_isspace(c)           isspace((int)c)
 #define lwip_isupper(c)           isupper((unsigned char)(c))
 #define lwip_tolower(c)           tolower((unsigned char)(c))
 #define lwip_toupper(c)           toupper((unsigned char)(c))
